+++ @Shared/UnityMod/changelog.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## [1.0.1]
+- changed methods for labels and locals
+- added method InsertJump
+
 ## [1.0.0]
 - first release